Respect expansion during root serialization
