+++ editor/src/document/document_file.rs
@@ -11,6 +11,7 @@ use crate::message_prelude::*;
 use crate::EditorError;
 
 use glam::{DAffine2, DVec2};
+use graphene::layers::Folder;
 use kurbo::PathSeg;
 use log::warn;
 use serde::{Deserialize, Serialize};
@@ -201,6 +202,61 @@ impl DocumentMessageHandler {
 	pub fn selected_layers(&self) -> impl Iterator<Item = &[LayerId]> {
 		self.layer_data.iter().filter_map(|(path, data)| data.selected.then(|| path.as_slice()))
 	}
+
+
+	fn serialize_structure(&self, folder: &Folder, structure: &mut Vec<u64>, data: &mut Vec<LayerId>, path: &mut Vec<LayerId>) {
+		let mut space = 0;
+		for (id, layer) in folder.layer_ids.iter().zip(folder.layers()) {
+			data.push(*id);
+			space += 1;
+			match layer.data {
+				LayerDataType::Shape(_) => (),
+				LayerDataType::Folder(ref folder) => {
+                    path.push(*id);
+                    if self.layerdata(path).expanded {
+                        structure.push(space);
+                        self.serialize_structure(folder, structure, data, path);
+                        space = 0;
+                    }
+                    path.pop();
+				}
+			}
+		}
+		structure.push(space | 1 << 63);
+	}
+
+	/// Serializes the layer structure into a compressed 1d structure
+	///
+	/// It is a string of numbers broken into three sections:
+	/// (4),(2,1,-2,-0),(16533113728871998040,3427872634365736244,18115028555707261608,15878401910454357952,449479075714955186) <- Example encoded data
+	/// L = 4 = structure.len()                                                                                                 <- First value in the encoding: L, the length of the structure section
+	/// structure = 2,1,-2,-0                                                                                                   <- Subsequent L values: structure section
+	/// data = 16533113728871998040,3427872634365736244,18115028555707261608,15878401910454357952,449479075714955186            <- Remaining values: data section (layer IDs)
+	///
+	/// The data section lists the layer IDs for all folders/layers in the tree as read from top to bottom.
+	/// The structure section lists signed numbers. The sign indicates a folder indentation change (+ is down a level, - is up a level).
+	/// the numbers in the structure block encode the indentation,
+	/// 2 mean read two element from the data section, then place a [
+	/// -x means read x elements from the data section and then insert a ]
+	///
+	/// 2     V 1  V -2  A -0 A
+	/// 16533113728871998040,3427872634365736244,  18115028555707261608, 15878401910454357952,449479075714955186
+	/// 16533113728871998040,3427872634365736244,[ 18115028555707261608,[15878401910454357952,449479075714955186]    ]
+	///
+	/// resulting layer panel:
+	/// 16533113728871998040
+	/// 3427872634365736244
+	/// [3427872634365736244,18115028555707261608]
+	/// [3427872634365736244,18115028555707261608,15878401910454357952]
+	/// [3427872634365736244,18115028555707261608,449479075714955186]
+	pub fn serialize_root(&self) -> Vec<u64> {
+		let (mut structure, mut data) = (vec![0], Vec::new());
+		self.serialize_structure(self.graphene_document.root.as_folder().unwrap(), &mut structure, &mut data, &mut vec![]);
+		structure[0] = structure.len() as u64 - 1;
+		structure.extend(data);
+		structure
+	}
+
 
 	/// Returns the paths to all layers in order, optionally including only selected or non-selected layers.
 	fn layers_sorted(&self, selected: Option<bool>) -> Vec<Vec<LayerId>> {
@@ -454,11 +510,8 @@ impl MessageHandler<DocumentMessage, &InputPreprocessor> for DocumentMessageHand
 			}
 			ToggleLayerExpansion(path) => {
 				self.layer_data(&path).expanded ^= true;
-				match self.layer_data(&path).expanded {
-					true => responses.push_back(FolderChanged(path.clone()).into()),
-					false => responses.push_back(DocumentStructureChanged.into()),
-				}
-				responses.extend(self.layer_panel_entry(path.clone()).ok().map(|data| FrontendMessage::UpdateLayer { path: path.into(), data }.into()));
+                responses.push_back(DocumentStructureChanged.into());
+                responses.push_back(LayerChanged(path).into())
 			}
 			SelectionChanged => {
 				// TODO: Hoist this duplicated code into wider system
@@ -530,7 +583,7 @@ impl MessageHandler<DocumentMessage, &InputPreprocessor> for DocumentMessageHand
 				responses.extend([LayerChanged(path).into(), DocumentStructureChanged.into()]);
 			}
 			DocumentStructureChanged => {
-				let data_buffer: RawBuffer = self.graphene_document.serialize_root().into();
+				let data_buffer: RawBuffer = self.serialize_root().into();
 				responses.push_back(FrontendMessage::DisplayFolderTreeStructure { data_buffer }.into())
 			}
 			LayerChanged(path) => {
